refactor(sdk-pv-tagging): rename Storage class to Abstract Storage
